support m4 quant

+++ vllm_ascend/quantization/w8a8_pdmix.py
@@ -53,7 +53,7 @@ class AscendW8A8PDMixLinearMethod(AscendW8A8DynamicLinearMethod):
         layer.weight_scale.data = torch.flatten(layer.weight_scale.data)
         layer.weight_offset.data = torch.flatten(layer.weight_offset.data)
         layer.weight_scale_fp32 = layer.weight_scale.data.to(torch.float32)
-
+        layer.bias.data = layer.bias.data.to(layer.weight_scale.data.dtype)
 
 class AscendW8A8PDMixFusedMoeMethod(AscendW8A8DynamicFusedMoEMethod):
